% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difn.R
\name{difn}
\alias{difn}
\title{Find differences}
\usage{
difn(x, n = 1)
}
\arguments{
\item{x}{a numeric vecor}

\item{n}{an integer, pos or neg}
}
\value{
numeric vector
}
\description{
Finds the difference between two vector components, n places apart.  Like diff(), but faster and pads missing values with 0.
}
\examples{
difn(c(1,2,4,7,12),n=2)

}

