% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmetric_check.R
\name{fitmetric_check}
\alias{fitmetric_check}
\title{Convert fitmetric arguments to matrices.}
\usage{
fitmetric_check(x, y)
}
\arguments{
\item{x}{a vector of class numeric or integer or a numeric matrix}

\item{y}{a vector of class numeric or integer or a numeric matrix}
}
\value{
a list containing x, y, N(number of rows), and dof(number of columns)
}
\description{
Checks classes of fitmetric arguments and converts them to matrices if they are not already.
}
\examples{
vlist <- fitmetric_check(runif(100000),runif(100000))
mlist <- fitmetric_check(matrix(rnorm(10000),ncol=10),matrix(rnorm(10000),ncol=10))

}

