% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KS_D.R
\name{KS_D}
\alias{KS_D}
\title{KS_D}
\usage{
KS_D(x, y)
}
\arguments{
\item{x}{a vector or matrix of real numbers}

\item{y}{a vector or matrix of real numbers}
}
\value{
a vector of numbers with length equal to the number of trials (N)
}
\description{
Calculates the two-sided KS statistic given vectors for x and y
}
\examples{
KS_D(c(1,2,3,4,5),c(1,2,3,4,4))
KS_D(matrix(runif(100),ncol=7),matrix(runif(100),ncol=7))

}

