% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{R2}
\alias{R2}
\title{R-squared}
\usage{
R2(x1, y1)
}
\arguments{
\item{x1}{a vector or matrix of real numbers}

\item{y1}{a vector or matrix of real numbers}
}
\value{
a vector of numbers with length equal to the number of trials (N)
}
\description{
Calculates R-squared given vectors for x and y
}
\examples{
R2(c(1,2,3,4,5),c(1,2,3,4,4))
R2(matrix(runif(10000),ncol=5),matrix(runif(10000),ncol=5))

}

