% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{rmse}
\usage{
rmse(obs, mdl)
}
\arguments{
\item{obs}{a vector or matrix of real numbers}

\item{mdl}{a vector or matrix of real numbers}
}
\value{
a vector of numbers with length equal to the number of trials (N)
}
\description{
Calculates rmse given vectors for y and y_pred
}
\examples{
rmse(c(2:20),c(2:20 + c(rep(0.1,5),rep(-0.2,10),rep(0.3,4))))
mata <- matrix(runif(10000),ncol=5)
matb <- matrix(runif(10000),ncol=5)
rmse(mata, matb)

}

