% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table_pctbyfuncs.R
\name{Table_pctbyfuncs}
\alias{Table_pctbyfuncs}
\title{Print boilerplate stats}
\usage{
Table_pctbyfuncs(dof, pctlist = c(0.9, 0.95, 0.99), ndecimals = 2,
  dist = rnorm, order = 5, fitmetriclist = c(R2, rmse), ...)
}
\arguments{
\item{dof}{an integer}

\item{pctlist}{a vector of real numbers between 0 and 1, minimum acceptable noise levels.}

\item{ndecimals}{an integer}

\item{dist}{a random number distribution function}

\item{order}{a real number less than 7}

\item{fitmetriclist}{a vector of goodness-of-fit functions}

\item{...}{any argument that functions within this routine might use}
}
\value{
data frame
}
\description{
Prints a standard set of stats for a given degree of freedom
}
\examples{
Table_pctbyfuncs(6, dist=rnorm, sd=0.3)

}

